mmc: queue: Drop call to blk_cleanup_queue()
commit 0d8587376878c8da5a4953025e621c5dceb8b76a
"mmc: core: Allocate per-request data using the block layer core"
added a call to blk_cleanup_queue() into the mmc_cleanup_queue()
call.

This might look intutitive (and that is why I made the mistake),
since the cleanup is called if we fail in mmc_init_queue(), but due
to the nature of the MMC stack we are cleaning up the queue
inside the stack once the queue has been started.

The actual call to blk_cleanup_queue() is in mmc_blk_put() in
block.c, calling it twice wreaks havoc.

Fixes: 0d8587376878 ("mmc: core: Allocate per-request data using the block layer core")
Reported-by: Heiner Kallweit <hkallweit1@gmail.com>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>