UPSTREAM: binder: remove binder_debug_no_lock mechanism
With the global lock, there was a mechanism to access
binder driver debugging information with the global
lock disabled to debug deadlocks or other issues.
This mechanism is rarely (if ever) used anymore
and wasn't needed during the development of
fine-grained locking in the binder driver.
Removing it.

BUG=chromium:755697
TEST=Run ARC++

Change-Id: I55a01dddaf75a19a364e732ea5d4bf9677990440
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 1cf29cf4295a)