cpu_pm: replace raw_notifier with atomic_notifier
This patch replaces a rwlock and raw notifier by an atomic notifier
protected by a spin_lock and RCU.

The main reason to do that is due to a 'scheduling while atomic' bug
in RT kernels on an ARM/ARM64 platforms.  On ARM/ARM64, rwlock
cpu_pm_notifier_lock in cpu_pm causes a potential schedule after IRQ
disable in idle call chain:

cpu_startup_entry
  cpu_idle_loop
    local_irq_disable()
    cpuidle_idle_call
      call_cpuidle
        cpuidle_enter
          cpuidle_enter_state
            ->enter :arm_enter_idle_state
              cpu_pm_enter/exit
                CPU_PM_CPU_IDLE_ENTER
                  read_lock(&cpu_pm_notifier_lock); <-- sleep in idle
                     __rt_spin_lock();
                        schedule();

The kernel splat is as follows:

 BUG: scheduling while atomic: swapper/1/0/0x00000002
 [<ffff0000086fae70>] arm_enter_idle_state+0x18/0x70
 Modules linked in:
 [<ffff0000086f9298>] cpuidle_enter_state+0xf0/0x218
 [<ffff0000086f93f8>] cpuidle_enter+0x18/0x20
 Preemption disabled at:
 [<ffff0000080fa234>] call_cpuidle+0x24/0x40
 [<ffff000008882fa4>] schedule_preempt_disabled+0x1c/0x28
 [<ffff0000080fa49c>] cpu_startup_entry+0x154/0x1f8
 [<ffff00000808e004>] secondary_start_kernel+0x15c/0x1a0

Daniel Lezcano said this notification was needed on ARM/ARM64
platforms.

Sebastian suggested using atomic_notifier instead of rwlock, which is
not only removing the sleeping in idle, but also getting better latency
improvement.

Signed-off-by: Alex Shi <alex.shi@linaro.org>
Acked-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>